==> scripts/23_comparing_corr_orth_vs_random/DE_random_vs_real_pentadactyl/01_anolis_mouse_alligator_vs_random.R <==
# This recovers all of the information needed for the comparison of DE genes to random genes. (a) gene lists, (b) TPMs, (c) fold change of TPMs
# TAS

library(qvalue)

# Call the differential expression gene lists.####
pval_mouse_12 <- read.delim("raw_data/edgeR_outputs/mouse_all_ONgene_pvals/mouse_edgeR_lrt.1vs2_all_ONgene.txt")
pval_mouse_23 <- read.delim("raw_data/edgeR_outputs/mouse_all_ONgene_pvals/mouse_edgeR_lrt.2vs3_all_ONgene.txt")
pval_mouse_34 <- read.delim("raw_data/edgeR_outputs/mouse_all_ONgene_pvals/mouse_edgeR_lrt.3vs4_all_ONgene.txt")
pval_mouse_45 <- read.delim("raw_data/edgeR_outputs/mouse_all_ONgene_pvals/mouse_edgeR_lrt.4vs5_all_ONgene.txt")

pval_alligator_12 <- read.delim("raw_data/edgeR_outputs/alligator_all_ONgene_pvals/alligator_edgeR_lrt.1vs2_all_ONgene.txt")
pval_alligator_23 <- read.delim("raw_data/edgeR_outputs/alligator_all_ONgene_pvals/alligator_edgeR_lrt.2vs3_all_ONgene.txt")
pval_alligator_34 <- read.delim("raw_data/edgeR_outputs/alligator_all_ONgene_pvals/alligator_edgeR_lrt.3vs4_all_ONgene.txt")
pval_alligator_45 <- read.delim("raw_data/edgeR_outputs/alligator_all_ONgene_pvals/alligator_edgeR_lrt.4vs5_all_ONgene.txt")

pval_anolis_12 <- read.delim("raw_data/edgeR_outputs/anolis_pairwise_PC1_all_ONgene_pvals/anolis_edgeR_lrt.pc.1vs2_all_ONgene.txt")
pval_anolis_23 <- read.delim("raw_data/edgeR_outputs/anolis_pairwise_PC1_all_ONgene_pvals/anolis_edgeR_lrt.pc.2vs3_all_ONgene.txt")
pval_anolis_34 <- read.delim("raw_data/edgeR_outputs/anolis_pairwise_PC1_all_ONgene_pvals/anolis_edgeR_lrt.pc.3vs4_all_ONgene.txt")
pval_anolis_45 <- read.delim("raw_data/edgeR_outputs/anolis_pairwise_PC1_all_ONgene_pvals/anolis_edgeR_lrt.pc.4vs5_all_ONgene.txt")

# Calculate q values
qval_mouse_12 <- pval_mouse_12[(pval_mouse_12$FDR < 0.05),]
qval_mouse_23 <- pval_mouse_23[(pval_mouse_23$FDR < 0.05),]
qval_mouse_34 <- pval_mouse_34[(pval_mouse_34$FDR < 0.05),]
qval_mouse_45 <- pval_mouse_45[(pval_mouse_45$FDR < 0.05),]

qval_alligator_12 <- pval_alligator_12[(pval_alligator_12$FDR < 0.05),]
qval_alligator_23 <- pval_alligator_23[(pval_alligator_23$FDR < 0.05),]
qval_alligator_34 <- pval_alligator_34[(pval_alligator_34$FDR < 0.05),]
qval_alligator_45 <- pval_alligator_45[(pval_alligator_45$FDR < 0.05),]

qval_anolis_12 <- pval_anolis_12[(pval_anolis_12$FDR < 0.05),]
qval_anolis_23 <- pval_anolis_23[(pval_anolis_23$FDR < 0.05),]
qval_anolis_34 <- pval_anolis_34[(pval_anolis_34$FDR < 0.05),]
qval_anolis_45 <- pval_anolis_45[(pval_anolis_45$FDR < 0.05),]

# Reduce to the othologous TFs for direct comparisons.
one2one_list <- read.delim("output_files/acammm_TFs_1121.txt") # this is generated by the script "mklist_one_to_one_TFS_ammmgg.R"

one_to_one_ac_12 <- one2one_list[one2one_list$ac_gene_id %in% rownames(qval_anolis_12),]
one_to_one_ac_23 <- one2one_list[one2one_list$ac_gene_id %in% rownames(qval_anolis_23),]
one_to_one_ac_34 <- one2one_list[one2one_list$ac_gene_id %in% rownames(qval_anolis_34),]
one_to_one_ac_45 <- one2one_list[one2one_list$ac_gene_id %in% rownames(qval_anolis_45),]

one_to_one_mm_12 <- one2one_list[one2one_list$mm_gene_id %in% rownames(qval_mouse_12),]
one_to_one_mm_23 <- one2one_list[one2one_list$mm_gene_id %in% rownames(qval_mouse_23),]
one_to_one_mm_34 <- one2one_list[one2one_list$mm_gene_id %in% rownames(qval_mouse_34),]
one_to_one_mm_45 <- one2one_list[one2one_list$mm_gene_id %in% rownames(qval_mouse_45),]

one_to_one_am_12 <- one2one_list[one2one_list$am_gene_id %in% rownames(qval_alligator_12),]
one_to_one_am_23 <- one2one_list[one2one_list$am_gene_id %in% rownames(qval_alligator_23),]
one_to_one_am_34 <- one2one_list[one2one_list$am_gene_id %in% rownames(qval_alligator_34),]
one_to_one_am_45 <- one2one_list[one2one_list$am_gene_id %in% rownames(qval_alligator_45),]


# Calculate TPMs ####
# mouse data ####
HTSeq_mm <- read.table("raw_data/mm_HTSeq.txt", sep="\t", header=T)
gene_length_mm <- read.delim ("raw_data/mm_gene_length_median.txt")
merged_mm <- merge(gene_length_mm, HTSeq_mm, by="mm_gene_id")
merged_mm <- merge(merged_mm, one2one_list, by="mm_gene_id")

# alligator data ####
HTSeq_am <- read.delim("raw_data/am_HTSeq.txt")
gene_length_am <- read.delim("raw_data/am_gene_length_median.txt")
am_gene_id <- read.delim("raw_data/am_gene_annotations.txt", na.strings = "")
am_gene_id <- am_gene_id[c("am_gene_id", "gene_symbol")]
merged_am <-merge(am_gene_id, HTSeq_am, by="am_gene_id")
merged_am <-merge(gene_length_am, merged_am, by="am_gene_id")
merged_am <-merge(one2one_list, merged_am, by="gene_symbol") 

# anolis data ####
HTSeq_ac <- read.delim("raw_data/ac_HTSeq.txt")
gene_length_ac <- read.delim("raw_data/ac_gene_length_median.txt")
merged_ac <- merge(gene_length_ac, HTSeq_ac, by="ac_gene_id")
merged_ac <- merge(merged_ac, one2one_list, by="ac_gene_id")
merged_ac <- merged_ac[order(merged_ac$gene_symbol),]


# calc TPM
fn.TPM_mm <-function(sample) {
  (sample*75*10^6)/(merged_mm$length_mm*sum((sample*75)/merged_mm$length_mm))
}
TPM_mm <- as.data.frame(sapply(merged_mm[3:(length(merged_mm)-3)], fn.TPM_mm))

fn.TPM_am <-function(sample) {
  (sample*75*10^6)/(merged_am$length_am*sum((sample*75)/merged_am$length_am))
}
TPM_am <- as.data.frame(sapply(merged_am[7:length(merged_am)], fn.TPM_am))

fn.TPM_ac <-function(sample) {
  (sample*75*10^6)/(merged_ac$length_ac*sum((sample*75)/merged_ac$length_ac))
}
TPM_ac <- as.data.frame(sapply(merged_ac[3:(length(merged_ac)-3)], fn.TPM_ac))


# tidy R
remove(am_gene_id, gene_length_am, gene_length_mm, ac_gene_id, one2one_list,
       pval_alligator_12, pval_alligator_23, pval_alligator_34, pval_alligator_45,
       pval_anolis_12, pval_anolis_23, pval_anolis_34, 
       pval_mouse_12, pval_mouse_23, pval_mouse_34, pval_mouse_45,
       qval_alligator_12, qval_alligator_23, qval_alligator_34, qval_alligator_45,
       qval_anolis_12, qval_anolis_23, qval_anolis_34, qval_anolis_45, 
       qval_mouse_12, qval_mouse_23, qval_mouse_34, qval_mouse_45,
       fn.TPM_am, fn.TPM_ac, fn.TPM_mm,
       HTSeq_am, HTSeq_ac, HTSeq_mm)


is.nan.data.frame <- function(x) # set function to replace NaNs in tables.
  do.call(cbind, lapply(x, is.nan))
names(merged_am)[names(merged_am) == 'am_gene_id.x'] <- 'am_gene_id'

# Calculate fold change for all genes, all species ####
# chicken #### 
TPM_ac_1 <- rowMeans(TPM_ac[1:3])
TPM_ac_2 <- rowMeans(TPM_ac[4:6])
TPM_ac_3 <- rowMeans(TPM_ac[7:9])
TPM_ac_4 <- rowMeans(TPM_ac[10:12])
TPM_ac_5 <- rowMeans(TPM_ac[13:15])
TPM_ac_avg <- data.frame("ac_gene_id"=as.vector(merged_ac$ac_gene_id),
                         "TPM_ac_1"=rowMeans(TPM_ac[1:3]),
                         "TPM_ac_2"=rowMeans(TPM_ac[4:6]),
                         "TPM_ac_3"=rowMeans(TPM_ac[7:9]),
                         "TPM_ac_4"=rowMeans(TPM_ac[10:12]),
                         "TPM_ac_5"=rowMeans(TPM_ac[13:15]))

gene_symbol <- merged_ac$gene_symbol
ac_gene_id <- merged_ac$ac_gene_id

tpm_fold_change_ac <- data.frame (gene_symbol, ac_gene_id,
                                  "change1v2"=(TPM_ac_2-TPM_ac_1) / ((TPM_ac_2+TPM_ac_1)/2),
                                  "change2v3"=(TPM_ac_3-TPM_ac_2) / ((TPM_ac_3+TPM_ac_2)/2),
                                  "change3v4"=(TPM_ac_4-TPM_ac_3) / ((TPM_ac_4+TPM_ac_3)/2),
                                  "change4v5"=(TPM_ac_5-TPM_ac_4) / ((TPM_ac_5+TPM_ac_4)/2) )
tpm_fold_change_ac[is.nan.data.frame(tpm_fold_change_ac)] <- 0


#  mouse ####
TPM_mm_1 <- rowMeans(TPM_mm[1:3])
TPM_mm_2 <- rowMeans(TPM_mm[4:6])
TPM_mm_3 <- rowMeans(TPM_mm[7:9])
TPM_mm_4 <- rowMeans(TPM_mm[10:12])
TPM_mm_5 <- rowMeans(TPM_mm[13:15])
TPM_mm_avg <- data.frame("mm_gene_id"=as.vector(merged_mm$mm_gene_id),
                         "TPM_mm_1"=rowMeans(TPM_mm[1:3]),
                         "TPM_mm_2"=rowMeans(TPM_mm[4:6]),
                         "TPM_mm_3"=rowMeans(TPM_mm[7:9]),
                         "TPM_mm_4"=rowMeans(TPM_mm[10:12]),
                         "TPM_mm_5"=rowMeans(TPM_mm[13:15]))

gene_symbol <- merged_mm$gene_symbol
mm_gene_id <- merged_mm$mm_gene_id

tpm_fold_change_mm <- data.frame (gene_symbol, mm_gene_id,
                                  "change1v2"=(TPM_mm_2-TPM_mm_1) / ((TPM_mm_2+TPM_mm_1)/2),
                                  "change2v3"=(TPM_mm_3-TPM_mm_2) / ((TPM_mm_3+TPM_mm_2)/2),
                                  "change3v4"=(TPM_mm_4-TPM_mm_3) / ((TPM_mm_4+TPM_mm_3)/2),
                                  "change4v5"=(TPM_mm_5-TPM_mm_4) / ((TPM_mm_5+TPM_mm_4)/2) )
tpm_fold_change_mm[is.nan.data.frame(tpm_fold_change_mm)] <- 0


# alligator ####
TPM_am_1_e <- rowMeans(TPM_am[1:3])
TPM_am_2_e <- rowMeans(TPM_am[4:6])
TPM_am_3_e <- rowMeans(TPM_am[7:8])
TPM_am_4_e <- rowMeans(TPM_am[9:11])
TPM_am_5_e <- rowMeans(TPM_am[12:14])

TPM_am_1_l <- rowMeans(TPM_am[15:17])
TPM_am_2_l <- rowMeans(TPM_am[18:20])
TPM_am_3_l <- rowMeans(TPM_am[21:23])
TPM_am_4_l <- rowMeans(TPM_am[24:26])
TPM_am_5_l <- rowMeans(TPM_am[27:29])

TPM_am_avg <- data.frame("am_gene_id"=as.vector(merged_am$am_gene_id),
                         "TPM_am_1_e"=rowMeans(TPM_am[1:3]), "TPM_am_2_e"=rowMeans(TPM_am[4:6]), "TPM_am_3_e"=rowMeans(TPM_am[7:8]), "TPM_am_4_e"=rowMeans(TPM_am[9:11]), "TPM_am_5_e"=rowMeans(TPM_am[12:14]),
                         "TPM_am_1_l"=rowMeans(TPM_am[15:17]), "TPM_am_2_l"=rowMeans(TPM_am[18:20]), "TPM_am_3_l"=rowMeans(TPM_am[21:23]), "TPM_am_4_l"=rowMeans(TPM_am[24:26]), "TPM_am_5_l"=rowMeans(TPM_am[27:29]))

tpm_fold_change_am <- data.frame ("gene_symbol"=merged_am$gene_symbol,
                                  "am_gene_id"=merged_am$am_gene_id,
                                  "change1v2_e"=(TPM_am_2_e-TPM_am_1_e) / ((TPM_am_2_e+TPM_am_1_e)/2),
                                  "change1v2_l"=(TPM_am_2_l-TPM_am_1_l) / ((TPM_am_2_l+TPM_am_1_l)/2),
                                  "change2v3_e"=(TPM_am_3_e-TPM_am_2_e) / ((TPM_am_3_e+TPM_am_2_e)/2),
                                  "change2v3_l"=(TPM_am_3_l-TPM_am_2_l) / ((TPM_am_3_l+TPM_am_2_l)/2),
                                  "change3v4_e"=(TPM_am_4_e-TPM_am_3_e) / ((TPM_am_4_e+TPM_am_3_e)/2),
                                  "change3v4_l"=(TPM_am_4_l-TPM_am_3_l) / ((TPM_am_4_l+TPM_am_3_l)/2),
                                  "change4v5_e"=(TPM_am_5_e-TPM_am_4_e) / ((TPM_am_5_e+TPM_am_4_e)/2),
                                  "change4v5_l"=(TPM_am_5_l-TPM_am_4_l) / ((TPM_am_5_l+TPM_am_4_l)/2) )
tpm_fold_change_am[is.nan.data.frame(tpm_fold_change_am)] <- 0
